requirements.txt
# =============================================================================
# OSRS Market Data Collector - Python Dependencies
# =============================================================================

# Database drivers
psycopg2-binary==2.9.9          # PostgreSQL/TimescaleDB driver with connection pooling

# HTTP requests
requests==2.31.0                # API client for Wiki RS API
urllib3==2.1.0                  # HTTP client library (required by requests)

# Timezone handling
pytz==2023.3.post1             # Timezone database and utilities

# Environment configuration
python-dotenv==1.0.0           # Load environment variables from .env file

# Logging and monitoring
structlog==23.2.0              # Structured logging for better log parsing

# Data validation
typing-extensions==4.9.0       # Extended type hints for Python 3.9+

# Signal handling and utilities
python-json-logger==2.0.7      # JSON formatter for structured logging

# =============================================================================
# Optional Dependencies (uncomment if needed)
# =============================================================================

# For advanced database ORM (if needed in future)
# SQLAlchemy==2.0.23

# For caching layer (if needed in future)
# redis==5.0.1

# For testing (development only)
# pytest==7.4.4
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0

# For type checking (development only)
# mypy==1.7.1
# types-requests==2.31.0.10
# types-psycopg2==2.9.21.19

# =============================================================================
